# backend/openapi/v1.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "CPX Secreto API",
        "version": "v1"
    },
    "paths": {
...
        "/api/v1/groups/{group}/editions/{edition}/conversations": {
            "get": {
                "tags": [
                    "Conversations"
                ],
                "operationId": "listConversations",
                "parameters": [
                    {
                        "name": "group",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "edition",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Edition and assignment conversations visible to the authenticated participant.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ConversationCollection"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Authentication is required.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "An active edition participant is required.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "The group or edition is not visible.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/groups/{group}/editions/{edition}/conversations/{conversation}/messages": {
            "get": {
                "tags": [
                    "Conversations"
                ],
                "operationId": "getConversationMessages",
                "parameters": [
                    {
                        "name": "group",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "edition",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "conversation",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The latest messages in a visible conversation.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ConversationThread"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Authentication is required.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Only conversation participants may read messages.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "The group, edition, or conversation is not visible.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Conversations"
                ],
                "operationId": "createMessage",
                "parameters": [
                    {
                        "name": "group",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "edition",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "conversation",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateMessageRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Message sent.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Message"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Authentication is required.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Only conversation participants may send messages.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "The group, edition, or conversation is not visible.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "The conversation is not writable in the edition state.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "The request payload is invalid.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Too many messages were sent.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/groups/{group}/editions/{edition}/conversations/{conversation}/read": {
            "put": {
                "tags": [
                    "Conversations"
                ],
                "operationId": "markConversationRead",
                "parameters": [
                    {
                        "name": "group",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "edition",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "conversation",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MarkConversationReadRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Conversation marked as read."
                    },
                    "401": {
                        "description": "Authentication is required.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Only conversation participants may update read state.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "The group, edition, or conversation is not visible.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ConversationCollection": {
                "required": [
                    "data"
                ],
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Conversation"
                        }
                    }
                },
                "type": "object"
            },
            "ConversationCounterpart": {
                "required": [
                    "displayName",
                    "anonymous"
                ],
                "properties": {
                    "displayName": {
                        "oneOf": [
                            {
                                "type": "string",
                                "example": "João"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "anonymous": {
                        "type": "boolean",
                        "example": false
                    }
                },
                "type": "object"
            },
            "Conversation": {
                "required": [
                    "id",
                    "type",
                    "role",
                    "counterpart",
                    "unreadCount",
                    "lastMessageAt",
                    "canSend"
                ],
                "properties": {
                    "counterpart": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/ConversationCounterpart"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "lastMessageAt": {
                        "oneOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "edition",
                            "assignment"
                        ]
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "member",
                            "giver",
                            "receiver"
                        ]
                    },
                    "unreadCount": {
                        "type": "integer",
                        "example": 2,
                        "minimum": 0
                    },
                    "canSend": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
            "ConversationThread": {
                "required": [
                    "conversation",
                    "messages"
                ],
                "properties": {
                    "conversation": {
                        "$ref": "#/components/schemas/Conversation"
                    },
                    "messages": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Message"
                        }
                    }
                },
                "type": "object"
            },
            "CreateMessageRequest": {
                "required": [
                    "body"
                ],
                "properties": {
                    "body": {
                        "type": "string",
                        "example": "Qual tamanho você prefere?",
                        "maxLength": 1000,
                        "minLength": 1
                    }
                },
                "type": "object"
            },
            "MarkConversationReadRequest": {
                "required": [
                    "messageId"
                ],
                "properties": {
                    "messageId": {
                        "type": [
                            "integer",
                            "null"
                        ],
                        "minimum": 1
                    }
                },
                "type": "object"
            },
            "MessageAuthor": {
                "required": [
                    "displayName",
                    "anonymous"
                ],
                "properties": {
                    "displayName": {
                        "example": "João",
                        "oneOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "anonymous": {
                        "type": "boolean",
                        "example": false
                    }
                },
                "type": "object"
            },
            "Message": {
                "required": [
                    "id",
                    "body",
                    "isMine",
                    "author",
                    "sentAt"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "body": {
                        "type": "string",
                        "example": "Qual tamanho você prefere?",
                        "maxLength": 1000
                    },
                    "isMine": {
                        "type": "boolean",
                        "example": true
                    },
                    "author": {
                        "$ref": "#/components/schemas/MessageAuthor"
                    },
                    "sentAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "Error": {
                "required": [
                    "message"
                ],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Não autenticado."
                    },
                    "errors": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}